Added the user data to the repository
Updated the bulkloader.yaml to upload users data, had to massaged the date data a bit to get it working, migrating from sqlserver the data exported contains the 9 digits of the millisec or nano sec at the end, which the script doesn't like. Had to force to shorten it to 3 digits.